requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Werkzeug==2.2.3
python-dotenv==1.0.0
Jinja2>=3.0
Pillow==10.0.0
requests==2.31.0
Flask-WTF==1.1.1
WTForms==3.0.1
google-generativeai==0.3.1
